feat: Improve Platform Admins page APIs (#787)
# feat: Improve Platform Admins page APIs

## Description
- Removes the `GET ListCandidates`
- Improves the `GET List` PlatformAdmins with Paging, Search and Filters
to support the UI
- Improves the Paging/Search/Filtering infrastructure 

## Related Issues
- closes https://github.com/endatix/endatix-private/issues/419

## Type of Change
Check all that apply.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Dependency update
- [x] Refactoring (non-breaking change which improves code quality)
- [ ] Other (please describe)

## Checklist
- My code follows the style guidelines of this project
- I have performed a self-review of my own code
- I have commented my code, particularly in hard-to-understand areas
- I have made corresponding changes to the documentation
- My changes generate no new warnings
- I have added tests that prove my fix is effective or that my feature
works
- New and existing tests pass locally with my changes
- Any dependent changes have been merged and published in downstream
modules

> [!NOTE]
> - [x] I confirm that my Pull Request follows all the checklist
requirements above

## Screenshots
If applicable, add screenshots to help explain your changes.

+++ src/Endatix.Core/Infrastructure/Result/IPagedData.cs
@@ -1,26 +1,12 @@
 namespace Endatix.Core.Infrastructure.Result;
 
 /// <summary>
-/// Interface for paged data.
+/// Common paging metadata for paged results.
 /// </summary>
 public interface IPagedData
 {
-    /// <summary>
-    /// The page number.
-    /// </summary>  
     int Page { get; }
-    /// <summary>
-    /// The page size.
-    /// </summary>
     int PageSize { get; }
-    
-    /// <summary>
-    /// The total number of records.
-    /// </summary>
-    long TotalRecords { get; }
-
-    /// <summary>
-    /// The total number of pages.
-    /// </summary>
-    long TotalPages { get; }
+    int TotalRecords { get; }
+    int TotalPages { get; }
 }
